# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Test Platform API",
    "description": "API documentation for the AI Test Platform to manage projects, pages, tests, logs, settings, API endpoints, AI models, DOM scanning, and test generation.",
    "version": "1.0.0"
  },
  "paths": {
    "/projects": {
      "get": {
        "summary": "List all projects",
        "description": "Returns a list of all projects with their associated pages, tests, and API endpoints (Phase 1.1).",
        "responses": {
          "200": {
            "description": "List of projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Project" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new project",
        "description": "Creates a new project with name and baseUrl (Phase 1.1).",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateProjectDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created project",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Project" }
              }
            }
          }
        }
      }
    },
    "/projects/{id}": {
      "get": {
        "summary": "Get a project by ID",
        "description": "Returns details of a specific project (Phase 1.1).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Project details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Project" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a project",
        "description": "Updates a project's name or baseUrl (Phase 1.1).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateProjectDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated project",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Project" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a project",
        "description": "Deletes a project and its associated data (Phase 1.1).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "204": { "description": "Project deleted" }
        }
      }
    },
...
    "/tests/{id}": {
      "get": {
        "summary": "Get a test by ID",
        "description": "Returns details of a specific test (Phase 1.3).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Test details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Test" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a test",
        "description": "Updates a test's framework, script, or status (Phase 1.3).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTestDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated test",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Test" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a test",
        "description": "Deletes a test (Phase 1.3).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "204": { "description": "Test deleted" }
        }
      }
    },
...
    "/settings": {
      "get": {
        "summary": "List all settings",
        "description": "Returns a list of all settings, optionally filtered by projectId (Phase 2.3, 4).",
        "parameters": [
          {
            "name": "projectId",
            "in": "query",
            "required": false,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of settings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Setting" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new setting",
        "description": "Creates a new setting for global or project-specific configuration (Phase 2.3, 4).",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateSettingDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created setting",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Setting" }
              }
            }
          }
        }
      }
    },
    "/settings/{id}": {
      "get": {
        "summary": "Get a setting by ID",
        "description": "Returns details of a specific setting (Phase 2.3, 4).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Setting details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Setting" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a setting",
        "description": "Updates a setting's key or value (Phase 2.3, 4).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateSettingDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated setting",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Setting" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a setting",
        "description": "Deletes a setting (Phase 2.3, 4).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "204": { "description": "Setting deleted" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Project": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "baseUrl": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "pages": { "type": "array", "items": { "$ref": "#/components/schemas/Page" } },
          "tests": { "type": "array", "items": { "$ref": "#/components/schemas/Test" } },
          "apiEndpoints": { "type": "array", "items": { "$ref": "#/components/schemas/ApiEndpoint" } }
        }
      },
      "CreateProjectDto": {
        "type": "object",
        "required": ["name", "baseUrl"],
        "properties": {
          "name": { "type": "string" },
          "baseUrl": { "type": "string" }
        }
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "baseUrl": { "type": "string" }
        }
      },
      "Page": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "url": { "type": "string" },
          "requirements": { "type": "string", "nullable": true },
          "project": { "$ref": "#/components/schemas/Project" }
        }
      },
...
      "Test": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "framework": { "type": "string" },
          "script": { "type": "string" },
          "status": { "type": "string" },
          "project": { "$ref": "#/components/schemas/Project" }
        }
      },
...
      "UpdateTestDto": {
        "type": "object",
        "properties": {
          "framework": { "type": "string" },
          "script": { "type": "string" },
          "status": { "type": "string" }
        }
      },
...
      "Setting": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "projectId": { "type": "integer", "nullable": true },
          "key": { "type": "string" },
          "value": { "type": "string" }
        }
      },
      "CreateSettingDto": {
        "type": "object",
        "required": ["key", "value"],
        "properties": {
          "projectId": { "type": "integer", "nullable": true },
          "key": { "type": "string" },
          "value": { "type": "string" }
        }
      },
      "UpdateSettingDto": {
        "type": "object",
        "properties": {
          "key": { "type": "string" },
          "value": { "type": "string" }
        }
      },
      "ApiEndpoint": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "method": { "type": "string" },
          "url": { "type": "string" },
          "group": { "type": "string", "nullable": true },
          "requirements": { "type": "string", "nullable": true },
          "collections": { "type": "array", "items": { "$ref": "#/components/schemas/ApiCollection" }, "nullable": true },
          "project": { "$ref": "#/components/schemas/Project" }
        }
      },
      "ApiCollection": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "projectId": { "type": "integer" },
          "endpoints": { "type": "array", "items": { "$ref": "#/components/schemas/ApiEndpoint" }, "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}